small reorganisation
